move index updater methods to own module
