+++ src/search/index.rs
@@ -1,4 +1,5 @@
 use crate::note::Note;
+use crate::search::index_updater::update_index_metadata;
 use crate::settings::SETTINGS;
 use crate::utils::ensure_directory_exists;
 use crate::walk::{has_extension, walk_files};
@@ -107,53 +108,4 @@ pub fn create_index_and_add_documents() -> tantivy::Result<()> {
     println!("Indexed {} documents", searcher.num_docs());
     update_index_metadata().expect("Error writing index metadata file");
     Ok(())
-}
-
-use std::fs;
-use std::io::Read;
-use std::process::{Command, Stdio};
-use std::time::SystemTime;
-
-fn get_metadata_path() -> PathBuf {
-    SETTINGS.get_cache_path().join("index_metadata.txt")
-}
-
-fn update_index_metadata() -> std::io::Result<()> {
-    let now = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
-    fs::write(get_metadata_path(), now.to_string())
-}
-
-pub fn index_needs_update() -> std::io::Result<bool> {
-    let metadata_path = get_metadata_path();
-    if !metadata_path.exists() {
-        return Ok(true);
-    }
-
-    let mut contents = String::new();
-    fs::File::open(metadata_path)?.read_to_string(&mut contents)?;
-    let last_update = contents.parse::<u64>().unwrap_or(0);
-    let now = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
-
-    Ok(now - last_update > 300) // 5 minutes
-}
-
-pub fn spawn_index_update() {
-    match Command::new(std::env::current_exe().unwrap())
-        .arg("index")
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-    {
-        Ok(child) => {
-            // Intentionally not waiting for the child process
-            std::mem::forget(child);
-        }
-        Err(e) => eprintln!("Failed to spawn indexing process: {}", e),
-    }
 }
