+++ src/search/mod.rs
@@ -1,4 +1,5 @@
 mod index;
+mod index_updater;
 mod query;
 
 pub use self::index::create_index_and_add_documents;
